Change random seed policy to per-event
The previous setting was a non-choice that guaranteed backward compatibility by effectively making the very seed choice random.
I am changing the policy to a "mostly reproducible" one (per event).
Once the event is generated, the processing is reproducible.
Generation itself (and in general whatever job starts from an EmptyEvent source module) still is not reproducible.