add initial version of EventBuilderInput_source
A first version of an EventBuilderInput source, which reads
artdaq ROOT files and populates art::Events. This will read
the artdaq files output by the SBND test stand DAQ, and turn
CAEN v1730 fragments in raw::AuxDetDigits in a new LArSoft-style
file. An analyzer module is included for testing/debugging, which
outputs the waveforms to a ROOT file as TGraphs.

To run (both the source and the analyzer), use the event builder
FHICL file:

    $ lar -c standard_eventbuilder.fcl some_artdaq_file.root